전역상태 변경중
