+++ src/App.js
@@ -2,12 +2,15 @@ import logo from "./logo.svg";
 import "./App.css";
 import Router from "./shared/Router";
 import GlobalStyle from "components/GlobalStyle";
+import LetterProvider from "context/LettersContext";
 
 function App() {
   return (
     <>
-      <GlobalStyle />
-      <Router />
+      <LetterProvider>
+        <GlobalStyle />
+        <Router />
+      </LetterProvider>
     </>
   );
 }
